update instructions to share projects on Discord
The previous instructions were not super clear to someone unfamiliar with Discord. It sounded like there was a separate DAGsHub Discord server specifically for Hacktoberfest 2021. Coupled with the broken invite link, this could imply that that Hacktoberfest 2021 server is offline.

I tried to clarify the instructions to match what I believe to be the most likely situation: people should share projects in the `#hacktoberfest` channel on the DAGsHub Discord server. I hope this is right.